+++ src/main/java/com/manovikas/storestock/dao/PriceQuantityRepository.java
@@ -9,12 +9,13 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.OptionalInt;
 @Repository
 public interface PriceQuantityRepository extends JpaRepository<PriceQuantity,Integer> {
-    Optional<Integer> findIdByBrandIdAndPriceItem(int brandId, Integer priceItem);
-    Optional<PriceQuantity> findByPriceItemAndBrandId(@Param("priceItem") Integer priceItem, @Param("brandId") Integer brandId);
+    Optional<Integer> findIdByBrandIdAndPriceItem(int brandId, BigDecimal priceItem);
+    Optional<PriceQuantity> findByPriceItemAndBrandId(@Param("priceItem") BigDecimal priceItem, @Param("brandId") Integer brandId);
     Optional<PriceQuantity> findByBrandIdAndId(int brandId, int priceQuantityId);
 
     @Modifying
